Создает js модули и импортирует их в точку входа
