+++ js/main.js
@@ -1,101 +1,8 @@
-// Написать необходимые функции для создания массива из 25 сгенерированных объектов. Каждый объект массива — описание фотографии, опубликованной пользователем.
-
-// Структура каждого объекта должна быть следующей:
-// 1 - id, число — идентификатор опубликованной фотографии. Это число от 1 до 25. Идентификаторы не должны повторяться.
-// 2 - url, строка — адрес картинки вида photos/{{i}}.jpg, где {{i}} — это число от 1 до 25. Адреса картинок не должны повторяться.
-// 3 - description, строка — описание фотографии. Описание придумайте самостоятельно.
-// 4 - likes, число — количество лайков, поставленных фотографии. Случайное число от 15 до 200.
-// 5 - comments, массив объектов — список комментариев, оставленных другими пользователями к этой фотографии.
-//     Количество комментариев к каждой фотографии — случайное число от 0 до 30.
-//     Все комментарии генерируются случайным образом. Пример описания объекта с комментарием:
-
-//     {
-//       id: 135,
-//       avatar: 'img/avatar-6.svg',
-//       message: 'В целом всё неплохо. Но не всё.',
-//       name: 'Артём',
-//     }
-
-// 6 - У каждого комментария есть идентификатор — id — любое число. Идентификаторы не должны повторяться.
-// 7 - Поле avatar — это строка, значение которой формируется по правилу img/avatar-{{случайное число от 1 до 6}}.svg. Аватарки подготовлены в директории img.
-// 8 - Для формирования текста комментария — message — вам необходимо взять одно или два случайных предложения из представленных ниже:
-
-//     Всё отлично!
-//     В целом всё неплохо. Но не всё.
-//     Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.
-//     Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.
-//     Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.
-//     Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!
-
-// 9 - Имена авторов также должны быть случайными. Набор имён для комментаторов составьте сами. Подставляйте случайное имя в поле name.
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const NAMES = [
-  'Драгомир',
-  'Ратибор',
-  'Добронрав',
-  'Ярослав',
-  'Велимудр',
-  'Зоран',
-  'Бажена',
-  'Любомила',
-  'Велислава',
-  'Забава',
-  'Деяна',
-  'Злата',
-];
-
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const PHOTO_ID_COUNT = 25;
-const MIN_LIKES = 15;
-const MAX_LIKES = 200;
-const MIN_COMMENTS = 0;
-const MAX_COMMENTS = 30;
-const MIN_ID_AVATAR = 1;
-const MAX_ID_AVATAR = 6;
-let idComment = 0;
-let idPhoto = 0;
-
-// Создание объекта с комментарием.
-
-const createComment = () => ({
-  id: ++idComment,
-  avatar: `img/avatar-${getRandomInteger(MIN_ID_AVATAR, MAX_ID_AVATAR)}.svg`,
-  message: `${MESSAGES[getRandomInteger(0, MESSAGES.length - 1)]}`,
-  name: `${NAMES[getRandomInteger(0, NAMES.length - 1)]}`,
-});
-
-// Создание объекта с описанием фотографии.
-
-const createDescriptionPhoto = () => ({
-  id: ++idPhoto,
-  url: `photos/${idPhoto}.jpg`,
-  description: `Фотография № ${idPhoto}`,
-  likes: getRandomInteger(MIN_LIKES, MAX_LIKES),
-  comments: Array.from({ length: getRandomInteger(MIN_COMMENTS, MAX_COMMENTS) }, createComment),
-});
-
-// Создание массива с описанием фото.
-
-const createPhotosArray = () => {
-  const photosArray = Array.from({ length: PHOTO_ID_COUNT }, createDescriptionPhoto);
-  return photosArray;
-};
-
-// createPhotosArray();
+import {createPhotosArray} from './data.js';
+import './upload-img.js';
+import './redact-img.js';
+import './send-data.js';
+import './view-img.js';
+import './filter-img.js';
 
 console.log(createPhotosArray());
